(CDAP-5172) Added a cdap-security-spi module which contains interfaces that security extensions will implement. The extensions will be loaded from a separate classloader, whose parent will contain classes from this module only.
