+++ src/main/java/co/cask/cdap/security/authorization/DatasetBasedAuthorizer.java
@@ -16,27 +16,99 @@
 
 package co.cask.cdap.security.authorization;
 
+import co.cask.cdap.api.data.DatasetInstantiationException;
+import co.cask.cdap.api.dataset.DatasetDefinition;
+import co.cask.cdap.api.dataset.DatasetProperties;
+import co.cask.cdap.api.dataset.table.Table;
+import co.cask.cdap.data.dataset.SystemDatasetInstantiator;
+import co.cask.cdap.data2.datafabric.dataset.DatasetsUtil;
 import co.cask.cdap.data2.dataset2.DatasetFramework;
+import co.cask.cdap.data2.dataset2.DatasetManagementException;
+import co.cask.cdap.data2.dataset2.MultiThreadDatasetCache;
 import co.cask.cdap.proto.id.EntityId;
+import co.cask.cdap.proto.id.NamespaceId;
 import co.cask.cdap.proto.security.Action;
 import co.cask.cdap.proto.security.Principal;
+import co.cask.cdap.security.spi.authorization.Authorizer;
+import co.cask.cdap.security.spi.authorization.UnauthorizedException;
+import co.cask.tephra.TransactionAware;
 import co.cask.tephra.TransactionExecutor;
 import co.cask.tephra.TransactionExecutorFactory;
 import co.cask.tephra.TransactionSystemClient;
+import com.google.common.base.Supplier;
+import com.google.common.base.Throwables;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Sets;
 import com.google.inject.Inject;
 
+import java.io.IOException;
 import java.util.Set;
 
 /**
  * {@link Authorizer} that uses a dataset to manage ACLs.
  */
-public class DatasetBasedAuthorizer extends DatasetBasedAuthorizationEnforcer implements Authorizer {
+public class DatasetBasedAuthorizer implements Authorizer {
+
+  private final Supplier<ACLDataset> acls;
+  private final Supplier<TransactionExecutor> aclsTx;
 
   @Inject
   DatasetBasedAuthorizer(final DatasetFramework dsFramework,
                          final TransactionExecutorFactory txExecutorFactory,
                          TransactionSystemClient txClient) {
-    super(dsFramework, txExecutorFactory, txClient);
+    final MultiThreadDatasetCache dsCache = new MultiThreadDatasetCache(
+      new SystemDatasetInstantiator(dsFramework, null, null), txClient,
+      new NamespaceId(ACLDataset.ID.getNamespace().getId()), null, null, null);
+    this.acls =
+      new Supplier<ACLDataset>() {
+        @Override
+        public ACLDataset get() {
+          Table table;
+          try {
+            table = dsCache.getDataset(ACLDataset.ID.getId());
+          } catch (DatasetInstantiationException e) {
+            try {
+              table = DatasetsUtil.getOrCreateDataset(
+                dsFramework, ACLDataset.ID, "table",
+                DatasetProperties.EMPTY, DatasetDefinition.NO_ARGUMENTS, null);
+            } catch (DatasetManagementException | IOException e1) {
+              throw Throwables.propagate(e1);
+            }
+          }
+          return new ACLDataset(table);
+        }
+      };
+    this.aclsTx = new Supplier<TransactionExecutor>() {
+      @Override
+      public TransactionExecutor get() {
+        return txExecutorFactory.createExecutor(ImmutableList.of((TransactionAware) acls.get()));
+      }
+    };
+  }
+
+  @Override
+  public void enforce(final EntityId entity, final Principal principal,
+                      final Action action) throws UnauthorizedException {
+    boolean allowed = aclsTx.get().executeUnchecked(new TransactionExecutor.Function<ACLDataset, Boolean>() {
+      @Override
+      public Boolean apply(ACLDataset acls) throws Exception {
+        Set<Action> unfulfilledActions = Sets.newHashSet(action);
+        for (EntityId current : entity.getHierarchy()) {
+          Set<Action> allowedActions = acls.search(current, principal);
+          if (allowedActions.contains(Action.ALL)) {
+            return true;
+          }
+          unfulfilledActions.removeAll(allowedActions);
+          if (unfulfilledActions.isEmpty()) {
+            return true;
+          }
+        }
+        return unfulfilledActions.isEmpty();
+      }
+    }, acls.get());
+    if (!allowed) {
+      throw new UnauthorizedException(principal, action, entity);
+    }
   }
 
   @Override
